Make the marketo JS file easier to replace.
The marketo-drupal.js file is expected to be updated for GDPR soon, so I
want to make updating this version from it easier. I’ve added comments 
to explain most of the intentional differences, and restored some of the
original differences that don’t really need to be different. (Even if 
that meant bringing in minor bugs / whitespace issues to this version.)

My ideal is for a diff between an updated marketo-drupal.js file and
this file to be understandable (once whitespace is ignored).